fix: properly handle case when controller method in inherited
Before library used `.bind` on method in order to keep `this` refers to controller instance,
this approach leads to some issues when decorator applied to field with lambda assigned, or when method is inherited.
This commit replaces this approach with lambda function to ensure method will be called even it inherited or lambda.